restructure code, adding pom.xml
please check the new structure!
