--- code/Command.java
@@ -1,3 +1,0 @@
-public interface Command {
-    void execute(String[] args);
-}
