--- code/SaveMode.java
@@ -1,5 +1,0 @@
-public enum SaveMode {
-    INPUTONLY,
-    ALL,
-    OUTPUTONLY
-}
