--- code/Steppable.java
@@ -1,7 +1,0 @@
-public interface Steppable {
-    /**
-     * Az interfészt implementáló osztályok
-     * definiálják.
-     */
-    void Step();
-}
