+++ drukmakori_sivatag/src/main/java/drukmakori_sivatag/Command.java
@@ -1,0 +1,5 @@
+package drukmakori_sivatag;
+
+public interface Command {
+    void execute(String[] args);
+}
