+++ drukmakori_sivatag/src/main/java/drukmakori_sivatag/SaveMode.java
@@ -1,0 +1,7 @@
+package drukmakori_sivatag;
+
+public enum SaveMode {
+    INPUTONLY,
+    ALL,
+    OUTPUTONLY
+}
