+++ drukmakori_sivatag/src/main/java/drukmakori_sivatag/Steppable.java
@@ -1,0 +1,9 @@
+package drukmakori_sivatag;
+
+public interface Steppable {
+    /**
+     * Az interfészt implementáló osztályok
+     * definiálják.
+     */
+    void Step();
+}
