+++ drukmakori_sivatag/src/main/java/drukmakori_sivatag/Tuple.java
@@ -1,0 +1,17 @@
+package drukmakori_sivatag;
+
+public class Tuple<X, Y> {
+    public X x;
+    public Y y;
+
+    public Tuple(X x, Y y) {
+        this.x = x;
+        this.y = y;
+    }
+    X ReturnFirst(){
+        return x;
+    }
+    Y ReturnSecond(){
+        return y;
+    }
+}
